+++ test/samples/loc.c
@@ -1,0 +1,9 @@
+#include <stdio.h>
+
+// empty line above
+int main() {
+	return 1;
+}
+/*
+total lines: 10, content lines: 8, code lines: 4
+*/